Backend/requirements.txt
Flask
python-dotenv
flask-restful
